autorizationRequirements separated into classes
